More import adjusting.
