+++ graphics/src/layer.rs
@@ -245,15 +245,3 @@ impl<'a> Layer<'a> {
         }
     }
 }
-
-/// Returns the number of total vertices & total indices of all [`Mesh`]es.
-pub fn attribute_count_of<'a>(meshes: &'a [Mesh<'a>]) -> (usize, usize) {
-    meshes
-        .iter()
-        .map(|Mesh { buffers, .. }| {
-            (buffers.vertices.len(), buffers.indices.len())
-        })
-        .fold((0, 0), |(total_v, total_i), (v, i)| {
-            (total_v + v, total_i + i)
-        })
-}
